requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
fonttools==4.37.3
idna==3.4
imagecodecs==2021.11.20
imageio==2.22.0
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.3
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.0
nbsphinx==0.8.9
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
opencv-contrib-python==4.5.3.56
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pooch==1.6.0
prompt-toolkit==3.0.31
psutil==5.9.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
pywin32==304
pyzmq==24.0.1
requests==2.28.1
scikit-image==0.17.1
scikit-learn==1.0.2
scikit-plot==0.3.7
scipy==1.7.3
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
xgboost==1.5.1
zipp==3.8.1
